implement a basic virtual machine (#2)
## Summary

Implement a small and simple yet powerful virtual machine. It's a stack
machine, with inspiration taken from WebAssembly.

## Details

The execution loop is implemented in the `vm` module, the opcode
specification can be found in `vmspec`.

The instruction set covers:
* integer/float arithmetic
* bit operations
* comparisons
* conversions
* memory operations (read, write, block copy)
* control flow
* basic stack management

In addition, the VM has built-in support for subroutines, to be able to
efficiently model more complex control-flow (such as `try`/`finally`).
Basic exception handling is also supported.

For interfacing with the host, ref-counted foreign `ref`s are and
imported procedures (which invoke a callback) are supported. Procedures
can be stubs, to facilitate just-in-time compilation.

All memory addresses are in a virtual address space, backed by a single
host memory region that can be grown as needed.

In general, the set of operations is meant to not make too many
assumptions, while still being efficient.