implement first iteration of the lowest-level language (#7)
## Summary

* add a specification, pass, test runner, and tests for the lowest-
  level language that immediately comes before VM bytecode
* add a generic `PackedTree` implementation for use as the passes' IR

## Details

### Lanuage Design

* the language provides simple abstractions over some VM concepts
* all global entity descriptions (types, globals, procs) are stored in
  separate sections, and they're referenced by index (for fast lookup)
* local continuations abstract over raw jumps
* the type system makes a distinction between signed and unsigned
  integers

In its current form, the language is likely too high-level, but it's a
good start to base further language development on.

### Pass Design

The pass currently operates on whole modules at a time (though there's
no multi-module support at the moment). Beyond some minor assertions,
syntax and proper typing is *not* checked by the pass -- the idea is
that during normal compilation, passes trust that their input is sane
and sound, with validation implemented separately.

### Storage

A simple generic `PackedTree` type (based on NimSkull's `MirTree`) is
used as the IR for the pass. The node kind enum to use is provided by
the `spec` module, alongside S-expression serialization/
deserialization.

Nodes only store the kind and a type-erased value. Extra information,
such as types, needs to use nodes. This keeps nodes small (currently
8-byte) and serialization easy.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>