ci: setup a job for checking and generating the languages
This ensures the grammar for each language is correct before merging.

No code is actually generated for the languages at the moment, but it
will be in the future. Koch and the passtool are quick enough to build,
so the executables aren't shared between the `binaries` job and
`languages` job.