implement a language with a `Record` and `Array` type (#15)
## Summary

The `L3` language extends `L1` with the structured `Record` and `Array`
types, replacing the unstructured `Blob` type. Path expressions, for
the structures, are also included.

## Details

`L3` is used as the name to leave space for another language in-between
`L1` and `L3`.

The structured type are only a thin abstraction over blob types at this
level, and they're still not allowed to be used with globals, procedure
parameters, or return types.

The lowering pass performs the lowering of `L3` into `L1` in a
straightforward manner, not yet employing any of the possible
optimizations, such as folding pointer arithmetic operations.