tester: implement `reject` feature
Summary
=======

* runners can now signal "test valid and there's an error" with exit
  code 2
* introduce the `reject` test spec key -- it defaults to 'false'

Details
=======

The tester now makes a distinction between:
1. valid tests that result in an error (e.g., semantic analysis for
   some piece of code failed)
2. invalid test (e.g., syntax error in test code), or runner crash

Aside from crashes, it's up to a runner to decide which category an
error falls into. Exit code 2 signals that there was an error that
falls into the first category.

The `reject` key specifies whether a "valid test but an error is
reported" result is expected.