vm: add `Unreachable` opcode
When executing an `Unreachable` instruction, the VM unconditionally
exits with an error of the L0 language.

The opcode is added for the convenience of the code generator / higher-
level passes: a procedure ending in an `Unreachable` statement now no
longer requires an (unnecessary) exit continuation in order to result
in well- formed bytecode.