source2il: implement support for returning tuples
The target IL doesn't support tuple types as return types, and thus
returning the value has to happen via an out parameter (i.e., a
pointer indirection).

For this to work, expression lists need to be used, but since the
target IL doesn't support those either, they too have to be lowered
immediately.

Finally, in order to use out parameter in `exprToIL`, due to a phase
ordering issue, the expression has to be analyzed twice when of tuple
type.