source2il: remove `Asgn` lowering
Summary
=======

Simplify `source2il` by removing the in-place assignment lowering for
complex types.

Details
=======

Assignments for complex types are supported in the `L10`, so the
lowering is no longer necessary.

The special-casing of complex types is removed for assignments. In
addition, wrapping an lvalue expression in a `Copy` operation is now
handled at where the expression is *embedded* (via the new `genUse`)
instead of where it's *produced*.