pass4: fix virtual-to-physical register mapping
The previous mapping only allocated physical registers (and marked them
live) on a block-by-block basis, which is a problem some graphs. For
example, a graph with three basic-blocks with virtual register pairs
`(B, A)`, `(B, C)`, and `(A, C)` exist would result in `A` and `C`
sharing the same physical register (when `A`, `B`, and `C` aren't used
anywhere else).

The mapping is changed to use simplified version of the contemporary
"live range" register allocation approach, which correctly handles
this case.