lang: `If` expression support (#46)
## Summary

Adds `(If <predicate> <body> <alternative>)` expression support, as
well as gated logic, via the `(If <predicated> <body>)` variant. 

## Details

The `body` only variants are always treated as `unit` type expressions,
while the exhaustive variants can be expressions of any variety so long
as the `body` and `alternative` branch types match (the same or
subtype).

Implemented `If`, along with a variety of tests for various failure
scenarios. Additionally, added the ability to provide extra message
info to the `unreachable` routine, this is useful for debugging.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>