spec: prepare for variables and parameters (#47)
## Summary

* subdivide expressions into l-value and r-value expressions
* add definitions for values, locations, and cells
* introduce the notion of linear, affine, and normal types

## Details

Some definitions and terminology was missing to be able to describe /
specify the semantics of variables and assignments. This commit aims to
extend the specification with the necessary parts.

As a preparation for future resource management, linear and affine
types are introduced already, though they're not of practical relevance
yet.