pass7: fix auto spawning
For auto-spawning locals to work with loops, empty entry blocks must
not be merged away, otherwise the generated code can be incorrect when
there's no BB preceding a loop join. The tests' expected outputs are
adjusted accordingly.

Allowing the use of un-initialized locals was a mistake, but it's
better to remedy that separately.