vm: fix `opcBitNot` validation (#57)
## Summary

* fix validation for `opcBitNot` expecting two inputs (instead of only
  one)
* support integers in hexadecimal form (both `0xn` and `#n`) with the
  assembler

## Details

The `opcBitNot` operation only needs a single stack operand, but the
validation required two. This is now fixed, and a test for `opcBitNot`
is added to make sure validation and execution of it works.

For convenience of writing the test, the assembler is extended to
support integers in hexadecimal form in all places where integer
numbers are expected.