add the `phy` CLI application (#61)
## Summary

Bundle the passes, `source2il`, the assembler, and the VM into a single
CLI application. The idea with it is to help with development, by
providing a convenient way to invoke the passes/VM/etc. and inspect the
(intermediate) output.

## Details

The `phy` program provides a small command line interface for
specifying the source and target language are, what IRs to print, and
what to do (only compile or also evaluate).

Much of the implementation is copied from the various test runner:
* `source2il` invocation and comes from the `expr` test runner
* rendering VM execution results without high-level type information
  is moved over from the `pass0` test runner (into `vmexec.nim`)

The program also provides the `--runner` option, which makes it
behave in a way such that it's suitable as being used as a test runner,
by lifting some input restrictions and altering the output formatting,
so that it can maybe replace the currently used custom test runners at
some point.