tester: implement support for external runners
If instead of a `runner.nim` there's a `runner.txt` in a test
directory, the `runner.txt` needs to provide a runner description (in
form of a shell command).

The placeholders `${file}` and `${args}` can be used in the command.
They're with the test file and test-provided arguments when running
a test. For convenience, placeholders also support a default value.