tests: replace some custom runners with `phy` (#63)
## Summary

* support using external runner programs with the `tester`
* replace the runners for the `passX` and `expr` tests with using `phy`

## Details

If a test directory contains a `runner.txt` instead of a `runner.nim`
file,  command provided in the text file is used for running tests
from that directory. The placeholder `${file}` and `${args}` are used
for specifying where the test file and test-provided arguments are to
be inserted.

To support external runners, the `tester` is restructured:
* the logic for running a single test is turned into a procedure
  (`runTest`)
* the tester no longer invokes itself -- `runTest` is called directly
  instead
* both tester-built and external runners use the `RunnerDesc` type
  for describing how they're invoked

In order to be used instead of the previous `expr` test runner, `phy`
now ignores VM modules without procedures when in `--runner` mode.

Finally, the old custom runners are removed and replaced with usage of
`phy`. The specs of some tests are adjusted to the slightly different
behaviour of `phy` (i.e., output formatting and arguments).