lang: add the logical `And` and `Or` operators (#66)
## Summary

Add the logical `And` and `Or` operators to the source language.

## Details

Both operators short-circuit evaluation of the second operand. For this
reason, dedicated expression kinds are used instead of built-in calls.

Since the second expression of both `And` and `Or` is evaluated
conditionally, it is analyzed within its own scope in order to prevent
declarations therein from being visible to the outside -- such locals
would be in an undefined state when the expression is not evaluated.

`And` is implemented as the statement version of `(If a b False)`; `Or`
is implemented as the statement version of `(If a True b)`.