source2il: fix code generation for usage of locals (#68)
## Summary

Fix a `source2il` translation bug causing internal syntax errors when
using locals as `not`, `And`, `Or`, `If` or `Return` operands.

## Details

In the L30 language, loading the underlying value stored in a local
requires using `Copy` (which is also enforced by the gramamr). This was
not considered everywhere in `source2il`, leading to the L30 output
having syntax errors.

The problematic translation is changed to use `genUse` instead of
`inline` or manually adding the node sequences. For convenience, a
`genUse` overload is added that takes an `Expr`.

A test is added for each case where compilation previously failed.