spec: disallow implicit moves for arguments
Assigning to the temporary as if done by `Asgn` means that moving from
l-value expressions is allowed, which makes it unnecessarily hard for
the programmer to reason about lifetimes (because every normal
parameter can result in a move).

Until a borrow checker is implemented, l-value arguments now always
require a copy.