lang: add `While` loops (#72)
## Summary

Add the `While` loop construct to the source language. A `While` loop
repeatedly evaluates a given unit expression as long as the given
condition evaluates to true.

## Details

A `While` loop is implemented as a `Loop` + conditional `Break` in the
target IL.

At the moment, a `(While (Ident "true") ...)` is always typed as
`void`, as - without a `Break`-like statement - control-flow can never
reach the code immediately following the loop. Due to the current lack
of early constant folding, detection of such loops needs to be syntax-
based, and thus more complex cases (such as `(Exprs (Ident "true"))`)
are *not* typed as `void`.