skully: override importc'ed procedures
The overrides are implemented in a module that's included in the
compilation via an implicit import. Upon discovering an importc'ed
procedure, the corresponding override is looked up, and on success, the
AST of the original procedure is replaced with that of the override.

Overrides only need to be provided for actually used (read, part of the
live procedure graph) procedures.