vm: use a simpler type representation (#75)
## Summary

Simplify the type representation used by the VM in order to reduce the
size of types (in memory) and cost of validation.

## Details

Instead of having a list of types and a list with parameters (for
signature types), theres now a list of signatures and a list of
parameter types. `TypeId`s can only refer to signature types now.

This has a multiple benefits:
* there's no more requirement that the first element in the type list
  is a `tkVoid` type
* the type environment takes up less space in memory (and on disk)
* looking up a parameter type requires one less indirection

The VM validation layer is updated accordingly. For the assembler, the
type directive is changed such that only signature types are allowed,
whose syntax is now `(T*) -> T`, which should make it clearer what the
syntax represents (compared to the previous `(Proc ...)`.

For the tests, the expected output of `pass0` tests as well as the `vm`
tests are updated to the changed assembler syntax.

There are some changes relevant to embedders of the VM:
* the `initTypeEnv` procedure is removed, as an empty `TypeEnv` is now
  valid
* `YieldReason.typ` is a `TypeKind` instead of `TypeId` now