lang0: implement foreign procedure support (#78)
## Summary

* add support for foreign procedures to `L0` through `L30` ILs. A
  foreign procedure is one whose implementation is provided separately  
* implement VM modules, which are used/needed for a post-code-
  generation link step

## Details

Foreign procedure are meant to be used for providing functionality or
host interaction (in case of the VM) that cannot be implemented
in-language.

They're defined via a `Foreign` definition in the procedure list of a
module in the `L0` through `L30` ILs. A `Foreign` definition specifies
a signature type and a name, the latter which identifies the foreign
procedure to the code generator / linker / etc. . Foreign procedures
are used / called like any other procedure.

For the VM backend, a foreign procedure is implemented via a VM
callback.

In order to handle the name to callback resolution, and as a
preparation for future multi-module support, the to-VM build process is
extend with a "link" step, now working as follows:
* `pass0` produces a `VmModule` for the input IL module (instead of
  appending the code directly to the `VmEnv`)
* all VM modules part of the build are linked via `vmmodules.link`
* all callbacks (provided by a `name -> callback` map) referenced by VM
  modules are added to the environment during linking

Since there's no multi-module support at the moment, there's no import/
export support yet. The implementation of `link` still supports more
than one module already, with in-code references to types, globals,
etc. being patched as needed.

The `host_impl` module provides the VM callbacks implementing the Phy
host API, which only consists of a single procedure used for testing
foreign procedure support.

Finally, the `tree_parser` is extended to support standalone string
literals (i.e., strings can appear in S-expression without an
accompanying node kind, but only when supported by the IR).

Two VM bugs are fixed:
* ignore stub and callback procedures when validating bodies (they
  previously resulted in false-positive errors)
* handle stub and callback procedures when disassembling, instead of
  crashing

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>