lang0: make globals explicitly typed
Summary
=======

Require definitions of global variables to also provide a type, instead
of just a literal value (from which the type is inferred).

Details
=======

Having an explicit type is a requirement for querying the type of
expressions with a global variable at the root, something which is not
implemented at the moment, but will be needed by future passes
(especially once the source language supports globals).

Having a dedicated node kind / subtree for a global definitions also
makes it easier/possible to extend the definition with additional
information (such as whether the global is mutable).