skully: fix globals having an incorrect offset
The entity nodes passed to the "discovered" event don't have a valid
type, leading to `VoidType` being passed to `newGlobal`. Since the void
type has an alignment value of 0 (which results in a 0xFFFFFFFF mask),
this zeroed the memory offset.

The type is now taken from the symbol of the global/constant, fixing
the bug.