implement a runner for skully-compiled programs
For now, the runner is integrated directly into skully, meaning that
compilation is always coupled with subsequent execution.

The `hostprocs` module provides the various foreign procedure
implementations. Not all of them adhere to the specification of the
respective C procedure and they're also not all safe (with regards to
sandboxing).