vm: fix illegal memory access not being detected
Summary
=======

Fix VM memory access checks not rejecting memory access when the
accessed memory region crosses the `uint64` upper limit, allowing for
read/write access to outside the guest memory.

Details
=======

The test in `checkmem` did not account for unsigned integer wrap
around. If `x` is greater than or equal to the start of the virtual
memory and `x + len` wraps around, then no error was reported if the
wrapped-around value was less than the virtual memory's upper limit.

Unsigned integer wrap around is now explicitly tested for, leading to
`checkmem` returning an error if detected.