implement the local-to-blob transformation pass
There are three blob-related passes, with the local-to-blob pass being
the last one. In theory, the stack allocation pass could turn the
locals that have their address taken into stack locations directly
(without the intermediate blob locals), but that would also require
the stack allocation pass to handle the block parameter fixup, which
is why a separate pass is used instead.