vm: fix incorrect stack state after `IndCall` (#82)
## Summary

* fix the operand stack being in an incorrect state after `IndCall`
* fix the assembler interpreting all numbers as hexadecimal numbers
* change the assembler to require a type name for `IndCall`

## Details

For convenience of the VM, the callee operand for `IndCall` must be on
top of the stack. It has to be popped before entering the procedure, so
that the arguments are on top of the stack, but it wasn't.

There were also two issues/bugs with the assembler that are now fixed:

* number parsing tries interpreting numbers as decimal first, so that
  numbers aren't always interpreted as hexadecimal (since `parseHex`
  also supports numbers without a prefix)

* `IndCall` parsing required a type ID rather than a type name, which
  is inconsistent with all other places expecting a type. A type name
  is now expected

Finally, tests for the `IndCall` operation are added to ensure that it
works correctly.