rework the passes and intermediate languages (#85)
## Summary

Rework and redesign the lower-level ILs, with a focus on simplifying
the passes, making them easier to change in the future. The passes are
also a lot faster now.

The `L25` and `L30` now also supports:
* aggregate parameters and return values
* locals that have their address taken
* alignment requirements for aggregate types

## Details

The `L30` and `L25` ILs stay largely the same, in terms of both grammar
and semantics, with the bulk of the changes being in the ILs sitting
between the `L25` and `L0`.

The focus with the redesign is on:
* adding support for aggregate parameters and return values
* making the passes and ILs easier to change in the future
* speeding up the passes to the point of being actually usable with
  real-world input
* reducing the amount of translation work for `pass0`

The major differences of the new ILs compared to their predecessors
are:
* no single-static assignment form is used, at least for now
* block parameters are only used by the entry block and `Except`
  blocks; `Continue`-with-value is removed
* the ILs below `L25` also use `Return`; there's no more exit
  continuation
* path expression use single-level `Path` subtrees, which are much
  easier to work with than the nested `Field`/`At` expressions
* there's an explicit `Union` type now
* all aggregate types have an explicitly specified alignment
  requirement

There are also some changes in the used terminology:
* `Continue` is now called `Goto`
* `SelectBool` is now called `Branch`
* continuations (`Continuation`) are now called basic-blocks (`Block`),
  which fits the construct better. This also frees up the term
  "continuation" for use in higher-level ILs

To make future reordering/reorganization of the passes easier, the new
passes use a descriptive names that don't include the name of the input
intermediate language.

Another big part of the rework is an improved `Changeset` API, which
now supports nested changes, as well as more convenient tree
construction using a small macro DSL. Internally, `Changeset` now uses
a mini-VM for copying the nodes from the input buffers to the output
buffer.

### Output Quality

* both register allocation and coalescing of locals is removed, meaning
  that the amount of stack space and locals used per procedure is
  usually much larger
* offsets for nested field / array element access are combined,
  resulting in shorter and more efficient bytecode

### Tests

The tests for the old passes are removed. Tests for the new passes are
added to sub-directories of the new `passes` test sub-directory. To
reduce both execution overhead and the amount of duplicated test
content, less test files are used for the new passes, with each
covering more ground than what was previously common.

The tester is changed to also look for tests in nested sub-directories.