vmmodules: implement exports
Both procedures and globals can be exported, though globals cannot be
imported at the moment. Callback procedure entries are generalized to
represent *imported procedures*, which resolve to both exported
procedures and host procedures (the former takes precedence). If an
imported procedure cannot be resolved, it is turned into a stub.