source2il: fix using call as field access operand (#93)
## Summary

Fix `source2il` crashing for call expressions used as the `FieldAccess`
operand.

## Details

Add the `inlineLvalue` procedure, which commits `L30` r-value
expressions to a temporary and then return the temporary. Since the
`FieldAccess` translation wants an `L30` l-value, it now uses the
new `inlineLvalue`.

With the removal of the aggregate return value lowering from
`source2il`, call expressions that return a tuple no longer use a
temporary. This resulted in `(Field (Call ...) ...)` `L30` trees,
which are not allowed by the `L30` grammar.