assembler: add the `.import` directive (#91)
## Summary

Add the `.import` directive to the assembler, which makes it possible
to express imported/foreign procedures in assembler code.

## Details

The import name uses double-quotes, even though the currently
supported character set doesn't require quotation. This is only
for reasons of style and readability, to highlight that the name
isn't a normal identifier.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>