vmmodules: support exported procs/globals (#94)
## Summary

* add support for exported procedures and globals to VM modules
* implement procedure import resolution for `vmmodules.link`

## Details

Procedures and global variables can now be marked as exported, by
adding them to the `export` table; the assembler provides this
functionality through the new `.export` directive. The same global
variable or procedure can be exported multiple times.

Exported procedures are used to resolve procedure imports, which
the `pkCallback` entries are repurposed to represent prior to
linking. The host procedure table is now only tried when a
procedure import doesn't resolve to an exported procedure; if
that fails too, the procedure is turned into a stub.

Global variables cannot be imported at the moment, but being able
to export them is still useful for looking them up from the
outside via an interface name.

Finally, since procedure stubs serve no purpose prior to linking, the
validator now reports an errors on  encountering them.

### Misc

The assembler was left in an invalid state when parsing the interface
name for the `.import` directive failed. This fixed by only modifying
the assembler after parsing the interface name.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>