vm: remove undefined behaviour (#96)
## Summary

Make the behaviour of all operations well-defined for all inputs and
fix the implementation of some operations exhibiting undefined
behaviour for some inputs.

## Details

The implementation of some opcodes used unchecked signed integer
arithmetic or conversions, thus triggering defects when building with
checks and undefined behaviour otherwise.

### Division By Zero

There's no sensible to value that can be returned for a integer
division by zero, so the `MulU`, `DivU`, `DivInt`, and `ModInt`
operations all trap (raise a run-time error) now.

### Float To Int Conversion

The `FloatToSInt` and `FloatToUInt` operations now perform a saturating
conversion, clamping the float value to the specified integer range.
NaNs are always converted to 0. The validator now also makes sure that
their immediate operand is a legal value.

### Negation

For `low(int64)`, `NegInt` and `DivInt` with a divisor of -1 overflow.
This could be resolved by trapping, but for consistency with `MulInt`
(which never traps on multiplication with -1), and to get around
branching control-flow for `Neg`, `NegInt` and `DivInt` with a divisor
of -1 now behave exactly like `MulInt` with a factor of -1.

### Address Computation

Computing the final address for `LdX` and `WrX` operations now uses
unsigned integer arithmetic, fixing the overflow issues the previous
code had. Behaviour in the non-overflow case doesn't change.

### Assembler

* the assembler now expects the name of a constant for `LdConst`
* fix integer-like values missing not being reported as an error
* fix parsing for typed value; the required space between the type and
  value wasn't skipped