spec: add a formal definition of the source language
Summary
=======

Replace the informal definition of the source language's syntax and
semantics with a formal one. This provides the basis for deriving
proofs of correctness in the future, as well as for proving that the
language is *type safe* (among other properties).

Details
=======

The formal definition attempts to capture the full current semantics of
the source language. To make the static and dynamic semantics simpler
to express, a *language core* is defined, which the source language
*desugars* to.

Same as before, the operational semantics of floating-point arithmetic
is not defined at the moment. In future work, it should be defined
according to the IEEE 754.2008 standard.

Where the formal definition (deliberately) differs from both the
informal one and currently implemented behaviour is integer overflow:
the new definition explicitly states that a panic is raised (i.e., the
program terminates), whereas the previous definition ignored the
possibility.

The old specification is kept around for now, as not all ideas in it
are formalized yet (because no implementation thereof exists at the
moment).